Minor compile fix.
Former-commit-id: 7f915a1949067846be46a373a73eadea836371a3

+++ alethzero/Transact.cpp
@@ -329,7 +329,7 @@ void Transact::rejigData()
 		er = ethereum()->call(s, value(), to, m_data, ethereum()->gasLimitRemaining(), gasPrice());
 	}
 	gasNeeded = (qint64)(er.gasUsed + er.gasRefunded);
-	htmlInfo = QString("<div class=\"info\"><span class=\"icon\">INFO</span> Gas required: %1 total = %2 base, %3 exec [%4 refunded later]</div>").arg(gasNeeded).arg(baseGas).arg(gasNeeded - baseGas).arg(er.gasRefunded) + htmlInfo;
+	htmlInfo = QString("<div class=\"info\"><span class=\"icon\">INFO</span> Gas required: %1 total = %2 base, %3 exec [%4 refunded later]</div>").arg(gasNeeded).arg(baseGas).arg(gasNeeded - baseGas).arg((qint64)er.gasRefunded) + htmlInfo;
 
 	if (er.excepted != TransactionException::None)
 	{
